+++ RobotRoot/src/main/java/com/RobotRoot/Services/ServiceEmpresa.java
@@ -3,49 +3,75 @@ package com.RobotRoot.Services;
 import com.RobotRoot.Entities.Empresa;
 import com.RobotRoot.Repositories.RepositorioEmpresa;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ReflectionUtils;
 
+import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
 @Service
 public class ServiceEmpresa {
-/*
-    Empresa ep1;
-    ServiceEmpleado se1 = new ServiceEmpleado();
-    public ServiceEmpresa (){
-        Empleado emple1 = this.se1.getEmpleado();
-        this.ep1=new Empresa("Cocalola", "Cr50#20-25", 3012251, "1235478-NIT", emple1);
-    }
 
-    public Empresa getEmpresa(){
-        return this.ep1;
-    }
-*/
     private RepositorioEmpresa repositorio;
     public ServiceEmpresa(RepositorioEmpresa repositorio){
         this.repositorio = repositorio;
     }
-    public List<Empresa> getRepositorio(){
+    public ArrayList<Empresa> listarEmpresas(){
+        return (ArrayList<Empresa>) this.repositorio.findAll();
+    }
+
+    public List<Empresa> listarEmpresa1(){
         return this.repositorio.findAll();
     }
 
-    public Empresa CrearRegistro(Empresa empresa1){
-        return this.repositorio.save(empresa1);
+    public Optional<Empresa> buscarEmpresa(Long id){
+        return this.repositorio.findById(id);
     }
 
-    public Empresa actualEmpresa (Long id, Empresa EmprActualiza){
-        Empresa empresaActual = repositorio.findById(id).orElseThrow();
-        empresaActual.setNombre(EmprActualiza.getNombre());
-        empresaActual.setDireccion(EmprActualiza.getDireccion());
-        empresaActual.setTelefono(EmprActualiza.getTelefono());
-        empresaActual.setNit(EmprActualiza.getNit());
-        return this.repositorio.save(empresaActual);
+    public Empresa buscarEmpresa1(Long id){
+        return this.repositorio.findById(id).get();
     }
 
-    public Empresa eliminEmpresa(Long id){
-        Empresa empresaActual = repositorio.findById(id).orElseThrow();
-        this.repositorio.deleteById(id);
-        return empresaActual;
 
+    public String crearEmpresa(Empresa empresa){
+        if(this.repositorio.findById(empresa.getId()).isEmpty()){
+            this.repositorio.save(empresa);
+            return "Se creó la empresa exitosamente";
+        }else{
+            return "No se creó la empresa.";
+        }
     }
 
+    public Empresa actualizarCampo(Long id, Map<Object,Object> empresaMapeo){
+        Empresa emp = this.repositorio.findById(id).get();
+
+        empresaMapeo.forEach((key,value)->{
+            Field campo = ReflectionUtils.findField(Empresa.class, (String) key);
+            campo.setAccessible(true);
+            ReflectionUtils.setField(campo, emp, value);
+        });
+
+        return this.repositorio.save(emp);
+    }
+
+    public String eliminarEmpresa(Long id){
+        if(this.repositorio.findById(id).isPresent()){
+            this.repositorio.deleteById(id);
+            return "Se eliminó la empresa exitosamente";
+        }else{
+            return "No se eliminó la empresa.";
+        }
+    }
+
+    public Empresa actualizarEmp(Long id, Empresa empresa){
+        Empresa emp = buscarEmpresa1(id);
+        emp.setDireccion(empresa.getDireccion());
+        emp.setNit(empresa.getNit());
+        emp.setNombre(empresa.getNombre());
+        emp.setTelefono(empresa.getTelefono());
+
+        return repositorio.save(emp);
+    }
 }
